+++ Common/src/Holodecoder.h
@@ -92,7 +92,7 @@ public:
     Holodecoder();
 	
 	virtual void init(void);
-    virtual void init(float width, float height);
+    virtual void init(int width, int height);
 	virtual void draw(void);
 	virtual void resize(int width, int height);
 	virtual void cameraSelectMode(int mode);
