+++ Desktop/src/Capture/ThreeFringe/ThreeFringeCapture.cpp
@@ -22,7 +22,7 @@ void ThreeFringeCapture::init()
   init(256,256);
 }
 
-void ThreeFringeCapture::init(float width, float height)
+void ThreeFringeCapture::init(int width, int height)
 {
   if(!m_hasBeenInit && width > 0 && height > 0)
   {
@@ -60,7 +60,7 @@ int ThreeFringeCapture::getHeight()
 	return m_height;
 }
 
-void ThreeFringeCapture::resizeInput(float width, float height)
+void ThreeFringeCapture::resizeInput(int width, int height)
 {
   //  Make sure that it has been initalized first.
   if(m_hasBeenInit)
@@ -88,10 +88,10 @@ void ThreeFringeCapture::resizeInput(float width, float height)
     m_imageProcessor.unbind();
 
     //  Send the new size to all of the shaders
-    m_phaseFilter.uniform("width", width);
-    m_phaseFilter.uniform("height", height);
-    m_normalCalculator.uniform("width", width);
-    m_normalCalculator.uniform("height", height);
+    m_phaseFilter.uniform("width", (float)width);
+    m_phaseFilter.uniform("height", (float)height);
+    m_normalCalculator.uniform("width", (float)width);
+    m_normalCalculator.uniform("height", (float)height);
 
     //  Resize the display mesh
     m_mesh = shared_ptr<TriMesh>(new TriMesh(width, height));
